requirements.txt
# Data Processing
pandas==2.0.3
numpy==1.24.3
apache-airflow==2.7.1
psycopg2-binary==2.9.6
pyspark==3.4.1
python-dotenv==1.0.0

# Data Quality
great-expectations==0.17.11

# Dashboard
streamlit==1.25.0
plotly==5.15.0

# Utils
requests==2.31.0
python-dateutil==2.8.2